Chess-Engine Pull Request  (#94)
* So we basically created a file in the folder api
called chess-engine.ts and moved the functions from
api.ts and redid them in chess-engine. After that,
we updated api.ts with new code we had in
chess-engine. Then we did the same in the common
folder. We moved the functions that was needed
from common to chessEngine.

* Remove comment

* So this is a small fix. I did what the comments
told me on the last pull. Here is the updated
version for chess-engine. For the undefined part.

* So this is an update on the chess-engine for
the undefined.

* I did the fen position as an argument and the
DEFAULT_POSITION. I did the Square instead of
string.I also made the getSquares function
that works similar to getMoves but maps the
results.

* Added aimove to chess-engine

* I did the makeAiMove(difficulty) in chess engine and did the getSquare in chess engine and chessboard-wrapper to check for the color

* Fixed test?

* Changed isWhite to side

* Formatted game

---------

Co-authored-by: Colin Wong <cwongmath@gmail.com>
Co-authored-by: Alex Kempen <alex.bookreader@gmail.com>