Added power sliders that dynamically update depending on the power levels of the robots.
This change updates the method of how we send
messages to the robots. We now update the state
with the different input methods, and if
the state changes, we send the message to the robot.
This allows us to ignore if the powerlevel is the same
as the previous one. Reduces message spam.